Move the new task into a separate class
